COMMIT6

+++ src/calculadora/Principal.java
@@ -7,32 +7,35 @@
  * *************************************************************************************************/
 
 package calculadora;
+
 import java.util.Scanner;
+
 public class Principal {
-	private static Scanner scanner; 
+	private static Scanner scanner;
+
 	public static void main(String[] args) {
-		scanner = new Scanner(System.in); 
-		System. out. print( "pon un numero  "  );
-		double n1 = scanner.nextInt();
-    	System. out. print( "pon el segundo numero  "  );
-    	double n2 = scanner.nextInt();
-		//double n1 = 10;
-		//double n2 = 5;
-		        
-		//suma
-		Suma sum = new Suma(n1,n2);
+		scanner = new Scanner(System.in);
+		System.out.print("pon un numero  ");
+		double n1 = scanner.nextDouble();
+		System.out.print("pon el segundo numero  ");
+		double n2 = scanner.nextDouble();
+		// double n1 = 10;
+		// double n2 = 5;
+
+		// suma
+		Suma sum = new Suma(n1, n2);
 		sum.mostrarResultado();
-		        
-		//resta
-		Resta res = new Resta(n1,n2);
+
+		// resta
+		Resta res = new Resta(n1, n2);
 		res.mostrarResultado();
-		        
-		//multiplicacion
-		Multiplicacion mul = new Multiplicacion(n1,n2);
+
+		// multiplicacion
+		Multiplicacion mul = new Multiplicacion(n1, n2);
 		mul.mostrarResultado();
-		        
-		//division
-		Divicion div = new Divicion(n1,n2);
+
+		// division
+		Divicion div = new Divicion(n1, n2);
 		div.mostrarResultado();
 
 	}
